更新部分を切り出し
