+++ src/components/applicationDetail/ApplicationContent.vue
@@ -1,5 +1,6 @@
 <script lang="ts" setup>
 import ApplicationAttachment from './ApplicationAttachment.vue'
+import { usePartitionInformation } from './composables/useApplicationInformation'
 import EditButton from '@/components/shared/EditButton.vue'
 import MarkdownIt from '@/components/shared/MarkdownIt.vue'
 import MarkdownTextarea from '@/components/shared/MarkdownTextarea.vue'
@@ -7,11 +8,9 @@ import SimpleButton from '@/components/shared/SimpleButton.vue'
 import UserIcon from '@/components/shared/UserIcon.vue'
 import { useApplication } from '@/features/application/composables'
 import type { ApplicationDetail } from '@/features/application/entities'
-import { useApplicationStore } from '@/features/application/store'
 import { useUserStore } from '@/features/user/store'
 import { formatDateAndTime } from '@/lib/date'
 import { ref } from 'vue'
-import { useToast } from 'vue-toastification'
 
 const props = defineProps<{
   application: ApplicationDetail
@@ -19,33 +18,17 @@ const props = defineProps<{
 
 const formattedDateAndTime = formatDateAndTime(props.application.createdAt)
 
-const toast = useToast()
 const { isApplicationCreator } = useApplication(props.application)
 
 const { me, userMap } = useUserStore()
-const { editApplication } = useApplicationStore()
 
 const hasAuthority = isApplicationCreator.value(me.value)
-const isEditMode = ref(false)
 const editedContent = ref(props.application.content)
-const toggleEditContent = () => {
-  if (isEditMode.value) {
-    editedContent.value = props.application.content
-  }
-  isEditMode.value = !isEditMode.value
-}
-const handleUpdateContent = async () => {
-  try {
-    await editApplication(props.application.id, {
-      ...props.application,
-      partition: props.application.partition.id,
-      content: editedContent.value
-    })
-    toast.success('更新しました')
-  } catch {
-    toast.error('更新に失敗しました')
-  }
-  isEditMode.value = false
+const { isEditMode, toggleEditContent, handleUpdateContent } =
+  usePartitionInformation(props.application, editedContent)
+
+const onUpdateClick = async () => {
+  await handleUpdateContent()
 }
 </script>
 
@@ -87,7 +70,7 @@ const handleUpdateContent = async () => {
         v-if="isEditMode"
         font-size="base"
         padding="sm"
-        @click="handleUpdateContent">
+        @click="onUpdateClick">
         完了
       </SimpleButton>
     </div>
